requirements.txt
streamlit==1.24.0
google-generativeai==0.3.1
python-dotenv==1.0.0
textstat==0.7.3 